# bpo-45636: Un-switch `BINARY_OP`